Resolve more Code Climate issues
