+++ src/io/exporters/FBXExporter.js
@@ -129,73 +129,19 @@ export default class FBXExporter extends Exporter {
         + `\t\tVersion: ${Version}\n`
         + `${FBXUtils.defaultProperties}`;
       /* Setting up vertices + indices */
-      const multiLayer = 0;
-      const multiTake = 1;
-      const shading = 'Y';
-      const culling = 'CullingOff';
-      const geometryVersion = 124;
-      /* About _correctArrayNotation: Float32Arrays will contains only Float32 numbers, which implies that it will be floating points with 17 numbers after point.
-      * We cannot (and it's logically incorrect) save all this information, so we convert this Float32Array into Array-like object with numbers with only 6 numbers after point
-      * Reminder - this is big memory loss (as we must save at one moment two arrays with similar information) */
-      const verticesIndices = `\t\tMultiLayer: ${multiLayer}\n`
-        + `\t\tMultiTake: ${multiTake}\n`
-        + `\t\tShading: ${shading}\n`
-        + `\t\tCulling: "${culling}"\n`
-        + `\t\tVertices: ${FBXUtils.correctArrayNotation(model.vertices)}\n\n`
-        + `\t\tPolygonVertexIndex: ${model.indices}\n\n`
-        + `\t\tGeometryVersion: ${geometryVersion}\n`;
+      const verticesIndices = FBXUtils.addVerticesIndices(model.vertices, model.indices);
       /* Setting up layers
       * Some positions are still not quite easy to understand, and as soon as it will not crash without them - they will be set with some default values */
-      const layerElementNormalNumber = 0; /* IDK what that is */
-      const layerElementNormalVersion = 101; /* IDK what version means */
-      const layerElementNormalName = ''; /* IDK what name means */
-      const layerElementNormal = `\t\tLayerElementNormal: ${layerElementNormalNumber} {\n`
-        + `\t\t\tVersion: ${layerElementNormalVersion}\n`
-        + `\t\t\tName: "${layerElementNormalName}"\n`
-        + '\t\t\tMappingInformationType: "ByVertice"\n' /* Mandatory for our Miew! Must not be changed */
-        + '\t\t\tReferenceInformationType: "Direct"\n' /* Mandatory for our Miew! Must not be changed */
-        + `\t\t\tNormals: ${FBXUtils.correctArrayNotation(model.normals)}\n`
-        + '\t\t}\n';
+      const layerElementNormal = FBXUtils.normalLayer(model.normals);
       /* next few layerElements are not in use, but we left it for maybe further compatibility */
       const layerElementSmoothing = '';
       const layerElementUV = '';
       const layerElementTexture = '';
       /* but colors and materials are actually in-use */
-      const layerElementColorNumber = 0; /* IDK what that is */
-      const layerElementColorVersion = 101; /* IDK what version means */
-      const layerElementColorName = ''; /* IDK what name means */
-      const layerElementColor = `\t\tLayerElementColor: ${layerElementColorNumber} {\n`
-        + `\t\t\tVersion: ${layerElementColorVersion}\n`
-        + `\t\t\tName: "${layerElementColorName}"\n`
-        + '\t\t\tMappingInformationType: "ByVertice"\n' /* Mandatory for our Miew! Must not be changed */
-        + '\t\t\tReferenceInformationType: "Direct"\n' /* Mandatory for our Miew! Must not be changed */
-        + `\t\t\tColors: ${FBXUtils.correctArrayNotation(model.colors)}\n`
-        + `\t\t\tColorIndex: ${[...Array(model.vertices.length / 3).keys()]}\n` /* As said - fastest and easiest way to produce [0, 1, .....] array */
-        + '\t\t}\n';
+      const layerElementColor = FBXUtils.colorLayer(model.colors);
+      const layerElementMaterial = FBXUtils.defaultMaterialLayer;
       /* Do we need automatically check and build this info? In Miew we always have these layers */
-      const layer = '\t\tLayer: 0 {\n'
-        + '\t\t\tVersion: 100\n'
-        + '\t\t\tLayerElement:  {\n'
-        + '\t\t\t\tType: "LayerElementNormal"\n'
-        + '\t\t\t\tTypedIndex: 0\n'
-        + '\t\t\t}\n'
-        + '\t\t\tLayerElement:  {\n'
-        + '\t\t\t\tType: "LayerElementColor"\n'
-        + '\t\t\t\tTypedIndex: 0\n'
-        + '\t\t\t}\n'
-        + '\t\t\tLayerElement:  {\n'
-        + '\t\t\t\tType: "LayerElementMaterial"\n'
-        + '\t\t\t\tTypedIndex: 0\n'
-        + '\t\t\t}\n'
-        + '\t\t}\n'
-        + '\t}\n';
-      const layerElementMaterial = '\t\tLayerElementMaterial: 0 {\n'
-        + '\t\t\tVersion: 101\n'
-        + '\t\t\tName: ""\n'
-        + '\t\t\tMappingInformationType: "AllSame"\n'
-        + '\t\t\tReferenceInformationType: "Direct"\n'
-        + '\t\t\tMaterials: 0\n'
-        + '\t\t}\n';
+      const layer = FBXUtils.defaultLayerBlock;
       // this._models[i] = null; /* That's free i guess? */
       const resultingLayer = layerElementNormal + layerElementSmoothing + layerElementUV + layerElementTexture + layerElementColor + layerElementMaterial + layer;
       allModels += (modelProperties + verticesIndices + resultingLayer);
@@ -412,7 +358,7 @@ export default class FBXExporter extends Exporter {
     const lVertices = _.cloneDeep(mesh.geometry.attributes.position.array);
     const lNormals = _.cloneDeep(mesh.geometry.attributes.normal.array);
     const lIndices = this._collectInstancedIndices(mesh, instanceIndex);
-    const lColors = this._collectInstancedColors(mesh, instanceIndex);
+    const lColors = FBXUtils.collectInstancedColors(mesh, instanceIndex);
     return [lVertices, lNormals, lIndices, lColors];
   }
 
@@ -445,65 +391,6 @@ export default class FBXExporter extends Exporter {
       }
     }
     return lIndices;
-  }
-
-  /**
-   * Collect and rework colors in big model notation.
-   * @param {object} mesh - mesh with instanced objects
-   * @param {Number} instanceIndex - number of instance in mesh
-   * @returns {Float32Array} array of gathered instanced colors
-   */
-  _collectInstancedColors(mesh, instanceIndex) {
-    const idxColors = (instanceIndex * 3); /* that's not magic. For 1st instance we must start from 0, for 2nd - from 3, etc */
-    const meshColor = mesh.geometry.attributes.color.array;
-    const meshAlphaColor = mesh.geometry.attributes.alphaColor.array;
-    const lAlphas = [meshAlphaColor[instanceIndex]];
-    const objectColor = FBXUtils.reworkColors([meshColor[idxColors], meshColor[idxColors + 1], meshColor[idxColors + 2]], lAlphas);
-    /* For FBX we need to clone color for every vertex */
-    const lColors = FBXUtils.cloneColors(mesh.geometry.attributes.position.array.length / 3, objectColor);
-    return lColors;
-  }
-
-  /**
-   * Calculate parameters for one cylinder in given mesh.
-   * @param {object} mesh - mesh with instanced objects
-   * @param {Number} instanceIndex - number of instance in mesh
-   * @returns {*[]} array of gathered transformations of vertices and normals
-   */
-  _calculateCylinderTransform(mesh, instanceIndex) {
-    /* Misc variables */
-    const matVector1 = mesh.geometry.attributes.matVector1.array;
-    const matVector2 = mesh.geometry.attributes.matVector2.array;
-    const matVector3 = mesh.geometry.attributes.matVector3.array;
-    /* Grab original vertices and normals */
-    const lVertices = _.cloneDeep(mesh.geometry.attributes.position.array);
-    const lNormals = _.cloneDeep(mesh.geometry.attributes.normal.array);
-    /* We have vertices for not transformed cylinder. Need to make it transformed */
-    const transformCylinder = new THREE.Matrix4();
-    const idxOffset = instanceIndex * 4;
-    transformCylinder.set(matVector1[idxOffset], matVector2[idxOffset], matVector3[idxOffset], 0,
-      matVector1[idxOffset + 1], matVector2[idxOffset + 1], matVector3[idxOffset + 1], 0,
-      matVector1[idxOffset + 2], matVector2[idxOffset + 2], matVector3[idxOffset + 2], 0,
-      matVector1[idxOffset + 3], matVector2[idxOffset + 3], matVector3[idxOffset + 3], 1);
-    /* For a reason we must perform transpose of that matrix */
-    transformCylinder.transpose();
-    const normVec = new THREE.Vector4();
-    const vertVec = new THREE.Vector4();
-    /* Applying offsets / transformation to every vertex */
-    for (let j = 0; j < lVertices.length; j += 3) {
-      vertVec.set(lVertices[j], lVertices[j + 1], lVertices[j + 2], 1);
-      vertVec.applyMatrix4(transformCylinder);
-      normVec.set(lNormals[j], lNormals[j + 1], lNormals[j + 2], 0.0);
-      normVec.applyMatrix4(transformCylinder);
-
-      lVertices[j] = vertVec.x;
-      lVertices[j + 1] = vertVec.y;
-      lVertices[j + 2] = vertVec.z;
-      lNormals[j] = normVec.x;
-      lNormals[j + 1] = normVec.y;
-      lNormals[j + 2] = normVec.z;
-    }
-    return [lVertices, lNormals];
   }
 
   /**
@@ -551,7 +438,7 @@ export default class FBXExporter extends Exporter {
     /* Main instances loop */
     for (let instanceIndex = 0; instanceIndex < numInstances - 1; ++instanceIndex) { /* Proceed every instance. Additional instance is strange. */
       /* Grab vertices and normals for transformed (scale, rotation, translation) cylinder */
-      const [lVertices, lNormals] = this._calculateCylinderTransform(mesh, instanceIndex);
+      const [lVertices, lNormals] = FBXUtils.calculateCylinderTransform(mesh, instanceIndex);
       const numVertices = lVertices.length / 3; /* That's the original number of vertices */
       /* Okay now vertices are reworked as we want them. Now it's time for implementing algorithm */
       /* Collect indices for given cylinder - remember: we will expand him later on */
